Add x-clacks-overhead
This adds a non-standard header, that was inspired by the Discworld series, and is used on other sites. There a series of towers are used to constantly re-transmit the memory of a lost son, throughout a network. Inspired by this many servers serve a similar header with memories of lost comrades (including Sir Terry Pratchett, the author).

Simon was a fan of the Discworld series as well, and I think this is a good way to remember him.

> In Sir Terry's novel "Going Postal", the story explains that the
> inventor of the Clacks - a man named Robert Dearheart, lost his only son
> in a suspicious workplace accident, and in order to keep the memory of
> his son alive, he transmitted his son's name as a special operational
> signal through the Clacks to forever preserve his memory:

https://xclacksoverhead.org/home/about