no redirect for /mnt/scratch in alphafold jobs
Hi @neoformit , we have a flag set for alphafold `tmp_dir: true` which should mean anything that would be written to /tmp will be written to <jwd>/tmp instead. It might be the case that this 'docker_volumes' entry has priority over that, and everything goes to /tmp which is not big enough in a handful of cases.