add tpv entry for morpheus
morpheus can run on pulsar. It is getting its cores/mem from the shared db as 1c/64gb. Because of the memory requirement it will almost always be better off on pulsar. Jobs like this with low CPU count and high memory are not handled well by the ranking function that sends the job to the destination with the lowest %CPUs allocated.