Change tpv entry mummer_nucmer: no singularity, no pulsar
I’ve added a long comment here too because this flies in the face of what we are trying to do. The latest version of this tool has issues running in the singularity container (previous versions were fine) and even though it’s a good pulsar tool, conda envs on pulsar are harder to maintain.