scrap embargo days and deny access to rstudio/jupyter for most users
This is sad, but the embargo time didn’t work. The bad user from recent times had had their account for months. There is another new account that is attempting to run the tools but can’t (because it is less than 7 days old).

Or should we just increase the embargo days and accept that occasionally a misuser will turn up?